ir,dse: Mege dse/promotion into clusters/algorithms
